prepare for rasa cli
