+++ rasa_core_sdk/constants.py
@@ -1,0 +1,1 @@
+DEFAULT_SERVER_PORT = 5055
