+++ tests/test_endpoint.py
@@ -7,16 +7,12 @@ import pytest
 import rasa_core_sdk.endpoint as ep
 
 
-def test_endpoint():
-    pass
-
-
 def test_arg_parser_actions_params_folder_style():
     parser = ep.create_argument_parser()
     args = ['--actions', 'actions/act']
 
     with pytest.raises(BaseException) as e:
-        cmdline_args = parser.parse_args(args)
+        parser.parse_args(args)
     if e is not None:
         assert True
     else:
